add FoodTypeTest, FoodRecordTest, FoodTest and FoodRecord

+++ src/main/java/model/UserProfile.java
@@ -2,12 +2,12 @@ package model;
 
 public class UserProfile {
     private double weight;
-    private double height;
+    private int height;
     private int age;
     private String gender;
     private double bmr; // Neues Feld für das Ergebnis
 
-    public UserProfile(double weight, double height, int age, String gender) {
+    public UserProfile(double weight, int height, int age, String gender) {
         this.weight = weight;
         this.height = height;
         this.age = age;
@@ -24,7 +24,7 @@ public class UserProfile {
     }
 
     public double getWeight() { return weight; }
-    public double getHeight() { return height; }
+    public int getHeight() { return height; }
     public int getAge() { return age; }
     public String getGender() { return gender; }
 }
